Add files via upload
This C++ program defines a function add that takes two integers as arguments and returns their sum. In the main function, it initializes two integers x and y with values 5 and 10, respectively, then calls the add function with x and y as arguments. The result is printed using cout with the message "Sum:". 
The output of this code will be:sum:15